.requirements.txt
Sphinx>=1.2.0,<1.3.0
flake8>=2.2.0,<2.3.0
flexmock>=0.9.0,<1.0.0
globusonline-transfer-api-client>=0.10.16,<0.11.0
mock>=1.0.0,<1.1.0
pytest>=2.6.0,<2.7.0
pytest-cov>=1.8.0,<1.9.0
python-coveralls>=2.4.0,<2.5.0
requests>=2.5.0,<2.6.0
subprocess32>=3.2.0,<3.3.0
